backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv==1.0.0
polars>=0.18.0
ell>=0.1.0
yfinance==0.2.36
pandas==2.2.1
polars==0.20.13
httpx==0.27.0
apscheduler==3.10.4
beautifulsoup4==4.12.3
openai==1.12.0